Updates YouTube Playlist Dialog Enhancer with search bar and clear button
Enhances the YouTube playlist dialog by adding a search bar for instant playlist filtering and a clear button to easily clear the search input. Improves the layout of the dialog by centering it on the screen and adding a clear button to the new playlist title field. Implements auto-population of channel names in the search bar and new playlist title field. Also includes various visual enhancements and bug fixes.